feat: Add public method for parsing cue payload
