+++ lib/text/cue.js
@@ -5,9 +5,13 @@
  */
 
 goog.provide('shaka.text.Cue');
-goog.provide('shaka.text.CueRegion');
 
+goog.require('shaka.log');
+goog.require('shaka.text.CueRegion');
 goog.require('shaka.util.ArrayUtils');
+goog.require('shaka.util.StringUtils');
+goog.require('shaka.util.TextParser');
+goog.require('shaka.util.TXml');
 
 
 /**
@@ -401,6 +405,397 @@ shaka.text.Cue = class {
 
     return true;
   }
+
+  /**
+   * Parses cue payload, searches for styling entities and, if needed,
+   * modifies original payload and creates nested cues to better represent
+   * styling found in payload. All changes are done in-place.
+   * @param {!shaka.text.Cue} cue
+   * @param {!Map<string, !shaka.text.Cue>=} styles
+   * @export
+   */
+  static parseCuePayload(cue, styles = new Map()) {
+    const StringUtils = shaka.util.StringUtils;
+    const TXml = shaka.util.TXml;
+    let payload = cue.payload;
+    if (!payload.includes('<')) {
+      cue.payload = StringUtils.htmlUnescape(payload);
+      return;
+    }
+    if (styles.size === 0) {
+      shaka.text.Cue.addDefaultTextColor(styles);
+    }
+    payload = shaka.text.Cue.replaceKaraokeStylePayload_(payload);
+    payload = shaka.text.Cue.replaceVoiceStylePayload_(payload);
+    payload = shaka.text.Cue.escapeInvalidChevrons_(payload);
+
+    cue.payload = '';
+
+    const xmlPayload = '<span>' + payload + '</span>';
+    let element;
+    try {
+      element = TXml.parseXmlString(xmlPayload, 'span');
+    } catch (e) {
+      shaka.log.warning('cue parse fail: ', e);
+    }
+    if (element) {
+      const childNodes = element.children;
+      if (childNodes.length == 1) {
+        const childNode = childNodes[0];
+        if (!TXml.isNode(childNode)) {
+          cue.payload = StringUtils.htmlUnescape(payload);
+          return;
+        }
+      }
+      for (const childNode of childNodes) {
+        shaka.text.Cue.generateCueFromElement_(childNode, cue, styles);
+      }
+    } else {
+      shaka.log.warning('The cue\'s markup could not be parsed: ', payload);
+      cue.payload = StringUtils.htmlUnescape(payload);
+    }
+  }
+
+  /**
+   * Add default color
+   *
+   * @param {!Map<string, !shaka.text.Cue>} styles
+   */
+  static addDefaultTextColor(styles) {
+    const textColor = shaka.text.Cue.defaultTextColor;
+    for (const [key, value] of Object.entries(textColor)) {
+      const cue = new shaka.text.Cue(0, 0, '');
+      cue.color = value;
+      styles.set('.' + key, cue);
+    }
+
+    const bgColor = shaka.text.Cue.defaultTextBackgroundColor;
+    for (const [key, value] of Object.entries(bgColor)) {
+      const cue = new shaka.text.Cue(0, 0, '');
+      cue.backgroundColor = value;
+      styles.set('.' + key, cue);
+    }
+  }
+
+  /**
+   * Converts karaoke style tag to be valid for xml parsing
+   * For example,
+   * input: Text <00:00:00.450> time <00:00:01.450> 1
+   * output: Text <div time="00:00:00.450"> time
+   *         <div time="00:00:01.450"> 1</div></div>
+   *
+   * @param {string} payload
+   * @return {string} processed payload
+   * @private
+   */
+  static replaceKaraokeStylePayload_(payload) {
+    const names = [];
+    let nameStart = -1;
+    for (let i = 0; i < payload.length; i++) {
+      if (payload[i] === '<') {
+        nameStart = i + 1;
+      } else if (payload[i] === '>') {
+        if (nameStart > 0) {
+          const name = payload.substr(nameStart, i - nameStart);
+          if (name.match(shaka.text.Cue.timeFormat_)) {
+            names.push(name);
+          }
+          nameStart = -1;
+        }
+      }
+    }
+    let newPayload = payload;
+    for (const name of names) {
+      const replaceTag = '<' + name + '>';
+      const startTag = '<div time="' + name + '">';
+      const endTag = '</div>';
+      newPayload = newPayload.replace(replaceTag, startTag);
+      newPayload += endTag;
+    }
+    return newPayload;
+  }
+
+  /**
+   * Converts voice style tag to be valid for xml parsing
+   * For example,
+   * input: <v Shaka>Test
+   * output: <v.voice-Shaka>Test</v.voice-Shaka>
+   *
+   * @param {string} payload
+   * @return {string} processed payload
+   * @private
+   */
+  static replaceVoiceStylePayload_(payload) {
+    const voiceTag = 'v';
+    const names = [];
+    let nameStart = -1;
+    let newPayload = '';
+    let hasVoiceEndTag = false;
+    for (let i = 0; i < payload.length; i++) {
+      // This condition is used to manage tags that have end tags.
+      if (payload[i] === '/') {
+        const end = payload.indexOf('>', i);
+        if (end === -1) {
+          return payload;
+        }
+        const tagEnd = payload.substring(i + 1, end);
+        if (!tagEnd || tagEnd != voiceTag) {
+          newPayload += payload[i];
+          continue;
+        }
+        hasVoiceEndTag = true;
+        let tagStart = null;
+        if (names.length) {
+          tagStart = names[names.length -1];
+        }
+        if (!tagStart) {
+          newPayload += payload[i];
+        } else if (tagStart === tagEnd) {
+          newPayload += '/' + tagEnd + '>';
+          i += tagEnd.length + 1;
+        } else {
+          if (!tagStart.startsWith(voiceTag)) {
+            newPayload += payload[i];
+            continue;
+          }
+          newPayload += '/' + tagStart + '>';
+          i += tagEnd.length + 1;
+        }
+      } else {
+        // Here we only want the tag name, not any other payload.
+        if (payload[i] === '<') {
+          nameStart = i + 1;
+          if (payload[nameStart] != voiceTag) {
+            nameStart = -1;
+          }
+        } else if (payload[i] === '>') {
+          if (nameStart > 0) {
+            names.push(payload.substr(nameStart, i - nameStart));
+            nameStart = -1;
+          }
+        }
+        newPayload += payload[i];
+      }
+    }
+    for (const name of names) {
+      const newName = name.replace(' ', '.voice-');
+      newPayload = newPayload.replace(`<${name}>`, `<${newName}>`);
+      newPayload = newPayload.replace(`</${name}>`, `</${newName}>`);
+      if (!hasVoiceEndTag) {
+        newPayload += `</${newName}>`;
+      }
+    }
+    return newPayload;
+  }
+
+  /**
+   * This method converts invalid > chevrons to HTML entities.
+   * It also removes < chevrons as per spec.
+   *
+   * @param {!string} input
+   * @return {string}
+   * @private
+   */
+  static escapeInvalidChevrons_(input) {
+    // Used to map HTML entities to characters.
+    const htmlEscapes = {
+      '< ': '',
+      ' >': ' &gt;',
+    };
+
+    const reEscapedHtml = /(< +>|<\s|\s>)/g;
+    const reHasEscapedHtml = RegExp(reEscapedHtml.source);
+    // This check is an optimization, since replace always makes a copy
+    if (input && reHasEscapedHtml.test(input)) {
+      return input.replace(reEscapedHtml, (entity) => {
+        return htmlEscapes[entity] || '';
+      });
+    }
+    return input || '';
+  }
+
+  /**
+   * @param {!shaka.extern.xml.Node} element
+   * @param {!shaka.text.Cue} rootCue
+   * @param {!Map<string, !shaka.text.Cue>} styles
+   * @private
+   */
+  static generateCueFromElement_(element, rootCue, styles) {
+    const TXml = shaka.util.TXml;
+    const nestedCue = rootCue.clone();
+    // We don't want propagate some properties.
+    nestedCue.nestedCues = [];
+    nestedCue.payload = '';
+    nestedCue.rubyTag = '';
+    // We don't want propagate some position settings
+    nestedCue.line = null;
+    nestedCue.region = new shaka.text.CueRegion();
+    nestedCue.position = null;
+    nestedCue.size = 0;
+
+    if (TXml.isNode(element)) {
+      const bold = shaka.text.Cue.fontWeight.BOLD;
+      const italic = shaka.text.Cue.fontStyle.ITALIC;
+      const underline = shaka.text.Cue.textDecoration.UNDERLINE;
+      const tags = element.tagName.split(/(?=[ .])+/g);
+      for (const tag of tags) {
+        let styleTag = tag;
+        // White blanks at start indicate that the style is a voice
+        if (styleTag.startsWith('.voice-')) {
+          const voice = styleTag.split('-').pop();
+          styleTag = `v[voice="${voice}"]`;
+          // The specification allows to have quotes and not, so we check to
+          // see which one is being used.
+          if (!styles.has(styleTag)) {
+            styleTag = `v[voice=${voice}]`;
+          }
+        }
+        if (styles.has(styleTag)) {
+          shaka.text.Cue.mergeStyle_(nestedCue, styles.get(styleTag));
+        }
+        switch (tag) {
+          case 'br': {
+            const lineBreakCue = shaka.text.Cue.lineBreak(
+                nestedCue.startTime, nestedCue.endTime);
+            rootCue.nestedCues.push(lineBreakCue);
+            return;
+          }
+          case 'b':
+            nestedCue.fontWeight = bold;
+            break;
+          case 'i':
+            nestedCue.fontStyle = italic;
+            break;
+          case 'u':
+            nestedCue.textDecoration.push(underline);
+            break;
+          case 'font': {
+            const color = element.attributes['color'];
+            if (color) {
+              nestedCue.color = color;
+            }
+            break;
+          }
+          case 'div': {
+            const time = element.attributes['time'];
+            if (!time) {
+              break;
+            }
+            const parser = new shaka.util.TextParser(time);
+            const cueTime = shaka.text.Cue.parseTime(parser);
+            if (cueTime) {
+              nestedCue.startTime = cueTime;
+            }
+            break;
+          }
+          case 'ruby':
+          case 'rp':
+          case 'rt':
+            nestedCue.rubyTag = tag;
+            break;
+          default:
+            break;
+        }
+      }
+    }
+
+    const isTextNode = (item) => TXml.isText(item);
+    const childNodes = element.children;
+    if (isTextNode(element) ||
+        (childNodes.length == 1 && isTextNode(childNodes[0]))) {
+      // Trailing line breaks may lost when convert cue to HTML tag
+      // Need to insert line break cue to preserve line breaks
+      const textArr = TXml.getTextContents(element).split('\n');
+      let isFirst = true;
+      for (const text of textArr) {
+        if (!isFirst) {
+          const lineBreakCue = shaka.text.Cue.lineBreak(
+              nestedCue.startTime, nestedCue.endTime);
+          rootCue.nestedCues.push(lineBreakCue);
+        }
+        if (text.length > 0) {
+          const textCue = nestedCue.clone();
+          textCue.payload = shaka.util.StringUtils.htmlUnescape(text);
+          rootCue.nestedCues.push(textCue);
+        }
+        isFirst = false;
+      }
+    } else {
+      rootCue.nestedCues.push(nestedCue);
+      for (const childNode of childNodes) {
+        shaka.text.Cue.generateCueFromElement_(childNode, nestedCue, styles);
+      }
+    }
+  }
+
+  /**
+   * Parses a WebVTT time from the given parser.
+   *
+   * @param {!shaka.util.TextParser} parser
+   * @return {?number}
+   */
+  static parseTime(parser) {
+    const results = parser.readRegex(shaka.text.Cue.timeFormat_);
+    if (results == null) {
+      return null;
+    }
+    // This capture is optional, but will still be in the array as undefined,
+    // in which case it is 0.
+    const hours = Number(results[1]) || 0;
+    const minutes = Number(results[2]);
+    const seconds = Number(results[3]);
+    const milliseconds = Number(results[4]);
+    if (minutes > 59 || seconds > 59) {
+      return null;
+    }
+
+    return (milliseconds / 1000) + seconds + (minutes * 60) + (hours * 3600);
+  }
+
+  /**
+   * Merges values created in parseStyle_
+   * @param {!shaka.text.Cue} cue
+   * @param {shaka.text.Cue} refCue
+   * @private
+   */
+  static mergeStyle_(cue, refCue) {
+    if (!refCue) {
+      return;
+    }
+
+    // Overwrites if new value string length > 0
+    cue.backgroundColor = shaka.text.Cue.getOrDefault_(
+        refCue.backgroundColor, cue.backgroundColor);
+    cue.color = shaka.text.Cue.getOrDefault_(
+        refCue.color, cue.color);
+    cue.fontFamily = shaka.text.Cue.getOrDefault_(
+        refCue.fontFamily, cue.fontFamily);
+    cue.fontSize = shaka.text.Cue.getOrDefault_(
+        refCue.fontSize, cue.fontSize);
+    cue.textShadow = shaka.text.Cue.getOrDefault_(
+        refCue.textShadow, cue.textShadow);
+
+    // Overwrite with new values as unable to determine
+    // if new value is set or not
+    cue.fontWeight = refCue.fontWeight;
+    cue.fontStyle = refCue.fontStyle;
+    cue.opacity = refCue.opacity;
+    cue.rubyTag = refCue.rubyTag;
+    cue.textCombineUpright = refCue.textCombineUpright;
+    cue.wrapLine = refCue.wrapLine;
+  }
+
+  /**
+   * @param {string} value
+   * @param {string} defaultValue
+   * @private
+   */
+  static getOrDefault_(value, defaultValue) {
+    if (value && value.length > 0) {
+      return value;
+    }
+    return defaultValue;
+  }
 };
 
 
@@ -554,118 +949,5 @@ shaka.text.Cue.textDecoration = {
   'OVERLINE': 'overline',
 };
 
-
-/**
- * @export
- */
-shaka.text.CueRegion = class {
-  /** */
-  constructor() {
-    const CueRegion = shaka.text.CueRegion;
-
-    /**
-     * Region identifier.
-     * @type {string}
-     * @export
-     */
-    this.id = '';
-
-    /**
-     * The X offset to start the rendering area in viewportAnchorUnits of the
-     * video width.
-     * @type {number}
-     * @export
-     */
-    this.viewportAnchorX = 0;
-
-    /**
-     * The X offset to start the rendering area in viewportAnchorUnits of the
-     * video height.
-     * @type {number}
-     * @export
-     */
-    this.viewportAnchorY = 0;
-
-    /**
-     * The X offset to start the rendering area in percentage (0-100) of this
-     * region width.
-     * @type {number}
-     * @export
-     */
-    this.regionAnchorX = 0;
-
-    /**
-     * The Y offset to start the rendering area in percentage (0-100) of the
-     * region height.
-     * @type {number}
-     * @export
-     */
-    this.regionAnchorY = 0;
-
-    /**
-     * The width of the rendering area in widthUnits.
-     * @type {number}
-     * @export
-     */
-    this.width = 100;
-
-    /**
-     * The width of the rendering area in heightUnits.
-     * @type {number}
-     * @export
-     */
-    this.height = 100;
-
-    /**
-     * The units (percentage, pixels or lines) the region height is in.
-     * @type {shaka.text.CueRegion.units}
-     * @export
-     */
-    this.heightUnits = CueRegion.units.PERCENTAGE;
-
-    /**
-     * The units (percentage or pixels) the region width is in.
-     * @type {shaka.text.CueRegion.units}
-     * @export
-     */
-    this.widthUnits = CueRegion.units.PERCENTAGE;
-
-    /**
-     * The units (percentage or pixels) the region viewportAnchors are in.
-     * @type {shaka.text.CueRegion.units}
-     * @export
-     */
-    this.viewportAnchorUnits = CueRegion.units.PERCENTAGE;
-
-    /**
-     * If scroll=UP, it means that cues in the region will be added to the
-     * bottom of the region and will push any already displayed cues in the
-     * region up.  Otherwise (scroll=NONE) cues will stay fixed at the location
-     * they were first painted in.
-     * @type {shaka.text.CueRegion.scrollMode}
-     * @export
-     */
-    this.scroll = CueRegion.scrollMode.NONE;
-  }
-};
-
-
-/**
- * @enum {number}
- * @export
- */
-shaka.text.CueRegion.units = {
-  'PX': 0,
-  'PERCENTAGE': 1,
-  'LINES': 2,
-};
-
-
-/**
- * @enum {string}
- * @export
- */
-shaka.text.CueRegion.scrollMode = {
-  'NONE': '',
-  'UP': 'up',
-};
+/** @private */
+shaka.text.Cue.timeFormat_ = /(?:(\d{1,}):)?(\d{2}):(\d{2})\.(\d{2,3})/g;
