+++ lib/text/mp4_vtt_parser.js
@@ -308,11 +308,8 @@ shaka.text.Mp4VttParser = class {
    * @private
    */
   static assembleCue_(payload, id, settings, startTime, endTime) {
-    const cue = new shaka.text.Cue(startTime, endTime, '');
-
-    /** @type {!Map.<string, shaka.text.Cue>} */
-    const styles = new Map();
-    shaka.text.VttTextParser.parseCueStyles(payload, cue, styles);
+    const cue = new shaka.text.Cue(startTime, endTime, payload);
+    shaka.text.Cue.parseCuePayload(cue);
 
     if (id) {
       cue.id = id;
