+++ lib/text/vtt_text_parser.js
@@ -15,7 +15,6 @@ goog.require('shaka.text.TextEngine');
 goog.require('shaka.util.Error');
 goog.require('shaka.util.StringUtils');
 goog.require('shaka.util.TextParser');
-goog.require('shaka.util.TXml');
 
 
 /**
@@ -106,9 +105,9 @@ shaka.text.VttTextParser = class {
       }
     }
 
-    /** @type {!Map.<string, shaka.text.Cue>} */
+    /** @type {!Map<string, !shaka.text.Cue>} */
     const styles = new Map();
-    VttTextParser.addDefaultTextColor_(styles);
+    shaka.text.Cue.addDefaultTextColor(styles);
 
     // Parse cues.
     const ret = [];
@@ -152,7 +151,7 @@ shaka.text.VttTextParser = class {
     }
 
     const parser = new shaka.util.TextParser(cueTimeMatch[1]);
-    const cueTime = shaka.text.VttTextParser.parseTime_(parser);
+    const cueTime = shaka.text.Cue.parseTime(parser);
     if (cueTime == null) {
       throw new shaka.util.Error(
           shaka.util.Error.Severity.CRITICAL,
@@ -164,28 +163,6 @@ shaka.text.VttTextParser = class {
     const mpegTimescale = shaka.text.VttTextParser.MPEG_TIMESCALE_;
 
     return time.periodStart + mpegTime / mpegTimescale - cueTime;
-  }
-
-  /**
-   * Add default color
-   *
-   * @param {!Map.<string, shaka.text.Cue>} styles
-   * @private
-   */
-  static addDefaultTextColor_(styles) {
-    const textColor = shaka.text.Cue.defaultTextColor;
-    for (const [key, value] of Object.entries(textColor)) {
-      const cue = new shaka.text.Cue(0, 0, '');
-      cue.color = value;
-      styles.set('.' + key, cue);
-    }
-
-    const bgColor = shaka.text.Cue.defaultTextBackgroundColor;
-    for (const [key, value] of Object.entries(bgColor)) {
-      const cue = new shaka.text.Cue(0, 0, '');
-      cue.backgroundColor = value;
-      styles.set('.' + key, cue);
-    }
   }
 
   /**
@@ -225,7 +202,7 @@ shaka.text.VttTextParser = class {
    * Parses a style block into a Cue object.
    *
    * @param {!Array.<string>} text
-   * @param {!Map.<string, shaka.text.Cue>} styles
+   * @param {!Map.<string, !shaka.text.Cue>} styles
    * @private
    */
   static parseStyle_(text, styles) {
@@ -372,7 +349,7 @@ shaka.text.VttTextParser = class {
    * @param {!Array.<string>} text
    * @param {number} timeOffset
    * @param {!Array.<!shaka.text.CueRegion>} regions
-   * @param {!Map.<string, shaka.text.Cue>} styles
+   * @param {!Map<string, !shaka.text.Cue>} styles
    * @return {shaka.text.Cue}
    * @private
    */
@@ -402,9 +379,9 @@ shaka.text.VttTextParser = class {
 
     // Parse the times.
     const parser = new shaka.util.TextParser(text[0]);
-    let start = VttTextParser.parseTime_(parser);
+    let start = shaka.text.Cue.parseTime(parser);
     const expect = parser.readRegex(/[ \t]+-->[ \t]+/g);
-    let end = VttTextParser.parseTime_(parser);
+    let end = shaka.text.Cue.parseTime(parser);
 
     if (start == null || expect == null || end == null) {
       shaka.log.alwaysWarn(
@@ -423,9 +400,9 @@ shaka.text.VttTextParser = class {
       cue = styles.get('global').clone();
       cue.startTime = start;
       cue.endTime = end;
-      cue.payload = '';
+      cue.payload = payload;
     } else {
-      cue = new shaka.text.Cue(start, end, '');
+      cue = new shaka.text.Cue(start, end, payload);
     }
 
     // Parse optional settings.
@@ -441,359 +418,12 @@ shaka.text.VttTextParser = class {
       word = parser.readWord();
     }
 
-    VttTextParser.parseCueStyles(payload, cue, styles);
+    shaka.text.Cue.parseCuePayload(cue, styles);
 
     if (id != null) {
       cue.id = id;
     }
     return cue;
-  }
-
-  /**
-   * Parses a WebVTT styles from the given payload.
-   *
-   * @param {string} payload
-   * @param {!shaka.text.Cue} rootCue
-   * @param {!Map.<string, shaka.text.Cue>} styles
-   */
-  static parseCueStyles(payload, rootCue, styles) {
-    const VttTextParser = shaka.text.VttTextParser;
-    const StringUtils = shaka.util.StringUtils;
-    const TXml = shaka.util.TXml;
-    // Optimization for unstyled payloads.
-    if (!payload.includes('<')) {
-      rootCue.payload = StringUtils.htmlUnescape(payload);
-      return;
-    }
-    if (styles.size === 0) {
-      VttTextParser.addDefaultTextColor_(styles);
-    }
-
-    payload = VttTextParser.replaceKaraokeStylePayload_(payload);
-    payload = VttTextParser.replaceVoiceStylePayload_(payload);
-    payload = VttTextParser.escapeInvalidChevrons_(payload);
-    const xmlPayload = '<span>' + payload + '</span>';
-    let element;
-    try {
-      element = TXml.parseXmlString(xmlPayload, 'span');
-    } catch (e) {
-      shaka.log.warning('cue parse fail: ', e);
-    }
-
-    if (element) {
-      const childNodes = element.children;
-      if (childNodes.length == 1) {
-        const childNode = childNodes[0];
-        if (!TXml.isNode(childNode)) {
-          rootCue.payload = StringUtils.htmlUnescape(payload);
-          return;
-        }
-      }
-      for (const childNode of childNodes) {
-        VttTextParser.generateCueFromElement_(childNode, rootCue, styles);
-      }
-    } else {
-      shaka.log.warning('The cue\'s markup could not be parsed: ', payload);
-      rootCue.payload = StringUtils.htmlUnescape(payload);
-    }
-  }
-
-  /**
-   * This method converts invalid > chevrons to HTML entities.
-   * It also removes < chevrons as per spec.
-   *
-   * @param {!string} input
-   * @return {string}
-   * @private
-   */
-  static escapeInvalidChevrons_(input) {
-    // Used to map HTML entities to characters.
-    const htmlEscapes = {
-      '< ': '',
-      ' >': ' &gt;',
-    };
-
-    const reEscapedHtml = /(< +>|<\s|\s>)/g;
-    const reHasEscapedHtml = RegExp(reEscapedHtml.source);
-    // This check is an optimization, since replace always makes a copy
-    if (input && reHasEscapedHtml.test(input)) {
-      return input.replace(reEscapedHtml, (entity) => {
-        return htmlEscapes[entity] || '';
-      });
-    }
-    return input || '';
-  }
-
-  /**
-   * Converts voice style tag to be valid for xml parsing
-   * For example,
-   * input: <v Shaka>Test
-   * output: <v.voice-Shaka>Test</v.voice-Shaka>
-   *
-   * @param {string} payload
-   * @return {string} processed payload
-   * @private
-   */
-  static replaceVoiceStylePayload_(payload) {
-    const voiceTag = 'v';
-    const names = [];
-    let nameStart = -1;
-    let newPayload = '';
-    let hasVoiceEndTag = false;
-    for (let i = 0; i < payload.length; i++) {
-      // This condition is used to manage tags that have end tags.
-      if (payload[i] === '/') {
-        const end = payload.indexOf('>', i);
-        if (end === -1) {
-          return payload;
-        }
-        const tagEnd = payload.substring(i + 1, end);
-        if (!tagEnd || tagEnd != voiceTag) {
-          newPayload += payload[i];
-          continue;
-        }
-        hasVoiceEndTag = true;
-        let tagStart = null;
-        if (names.length) {
-          tagStart = names[names.length -1];
-        }
-        if (!tagStart) {
-          newPayload += payload[i];
-        } else if (tagStart === tagEnd) {
-          newPayload += '/' + tagEnd + '>';
-          i += tagEnd.length + 1;
-        } else {
-          if (!tagStart.startsWith(voiceTag)) {
-            newPayload += payload[i];
-            continue;
-          }
-          newPayload += '/' + tagStart + '>';
-          i += tagEnd.length + 1;
-        }
-      } else {
-        // Here we only want the tag name, not any other payload.
-        if (payload[i] === '<') {
-          nameStart = i + 1;
-          if (payload[nameStart] != voiceTag) {
-            nameStart = -1;
-          }
-        } else if (payload[i] === '>') {
-          if (nameStart > 0) {
-            names.push(payload.substr(nameStart, i - nameStart));
-            nameStart = -1;
-          }
-        }
-        newPayload += payload[i];
-      }
-    }
-    for (const name of names) {
-      const newName = name.replace(' ', '.voice-');
-      newPayload = newPayload.replace(`<${name}>`, `<${newName}>`);
-      newPayload = newPayload.replace(`</${name}>`, `</${newName}>`);
-      if (!hasVoiceEndTag) {
-        newPayload += `</${newName}>`;
-      }
-    }
-    return newPayload;
-  }
-
-  /**
-   * Converts karaoke style tag to be valid for xml parsing
-   * For example,
-   * input: Text <00:00:00.450> time <00:00:01.450> 1
-   * output: Text <div time="00:00:00.450"> time
-   *         <div time="00:00:01.450"> 1</div></div>
-   *
-   * @param {string} payload
-   * @return {string} processed payload
-   * @private
-   */
-  static replaceKaraokeStylePayload_(payload) {
-    const names = [];
-    let nameStart = -1;
-    for (let i = 0; i < payload.length; i++) {
-      if (payload[i] === '<') {
-        nameStart = i + 1;
-      } else if (payload[i] === '>') {
-        if (nameStart > 0) {
-          const name = payload.substr(nameStart, i - nameStart);
-          if (name.match(shaka.text.VttTextParser.timeFormat_)) {
-            names.push(name);
-          }
-          nameStart = -1;
-        }
-      }
-    }
-    let newPayload = payload;
-    for (const name of names) {
-      const replaceTag = '<' + name + '>';
-      const startTag = '<div time="' + name + '">';
-      const endTag = '</div>';
-      newPayload = newPayload.replace(replaceTag, startTag);
-      newPayload += endTag;
-    }
-    return newPayload;
-  }
-
-  /**
-   * @param {string} value
-   * @param {string} defaultValue
-   * @private
-   */
-  static getOrDefault_(value, defaultValue) {
-    if (value && value.length > 0) {
-      return value;
-    }
-    return defaultValue;
-  }
-
-  /**
-   * Merges values created in parseStyle_
-   * @param {!shaka.text.Cue} cue
-   * @param {shaka.text.Cue} refCue
-   * @private
-   */
-  static mergeStyle_(cue, refCue) {
-    if (!refCue) {
-      return;
-    }
-
-    const VttTextParser = shaka.text.VttTextParser;
-    // Overwrites if new value string length > 0
-    cue.backgroundColor = VttTextParser.getOrDefault_(
-        refCue.backgroundColor, cue.backgroundColor);
-    cue.color = VttTextParser.getOrDefault_(
-        refCue.color, cue.color);
-    cue.fontFamily = VttTextParser.getOrDefault_(
-        refCue.fontFamily, cue.fontFamily);
-    cue.fontSize = VttTextParser.getOrDefault_(
-        refCue.fontSize, cue.fontSize);
-    cue.textShadow = VttTextParser.getOrDefault_(
-        refCue.textShadow, cue.textShadow);
-
-    // Overwrite with new values as unable to determine
-    // if new value is set or not
-    cue.fontWeight = refCue.fontWeight;
-    cue.fontStyle = refCue.fontStyle;
-    cue.opacity = refCue.opacity;
-    cue.rubyTag = refCue.rubyTag;
-    cue.textCombineUpright = refCue.textCombineUpright;
-    cue.wrapLine = refCue.wrapLine;
-  }
-
-  /**
-   * @param {!shaka.extern.xml.Node} element
-   * @param {!shaka.text.Cue} rootCue
-   * @param {!Map.<string, shaka.text.Cue>} styles
-   * @private
-   */
-  static generateCueFromElement_(element, rootCue, styles) {
-    const VttTextParser = shaka.text.VttTextParser;
-    const TXml = shaka.util.TXml;
-    const nestedCue = rootCue.clone();
-    // We don't want propagate some properties.
-    nestedCue.nestedCues = [];
-    nestedCue.payload = '';
-    nestedCue.rubyTag = '';
-    // We don't want propagate some position settings
-    nestedCue.line = null;
-    nestedCue.region = new shaka.text.CueRegion();
-    nestedCue.position = null;
-    nestedCue.size = 0;
-
-    if (shaka.util.TXml.isNode(element)) {
-      const bold = shaka.text.Cue.fontWeight.BOLD;
-      const italic = shaka.text.Cue.fontStyle.ITALIC;
-      const underline = shaka.text.Cue.textDecoration.UNDERLINE;
-      const tags = element.tagName.split(/(?=[ .])+/g);
-      for (const tag of tags) {
-        let styleTag = tag;
-        // White blanks at start indicate that the style is a voice
-        if (styleTag.startsWith('.voice-')) {
-          const voice = styleTag.split('-').pop();
-          styleTag = `v[voice="${voice}"]`;
-          // The specification allows to have quotes and not, so we check to
-          // see which one is being used.
-          if (!styles.has(styleTag)) {
-            styleTag = `v[voice=${voice}]`;
-          }
-        }
-        if (styles.has(styleTag)) {
-          VttTextParser.mergeStyle_(nestedCue, styles.get(styleTag));
-        }
-        switch (tag) {
-          case 'br': {
-            const lineBreakCue = shaka.text.Cue.lineBreak(
-                nestedCue.startTime, nestedCue.endTime);
-            rootCue.nestedCues.push(lineBreakCue);
-            return;
-          }
-          case 'b':
-            nestedCue.fontWeight = bold;
-            break;
-          case 'i':
-            nestedCue.fontStyle = italic;
-            break;
-          case 'u':
-            nestedCue.textDecoration.push(underline);
-            break;
-          case 'font': {
-            const color = element.attributes['color'];
-            if (color) {
-              nestedCue.color = color;
-            }
-            break;
-          }
-          case 'div': {
-            const time = element.attributes['time'];
-            if (!time) {
-              break;
-            }
-            const parser = new shaka.util.TextParser(time);
-            const cueTime = shaka.text.VttTextParser.parseTime_(parser);
-            if (cueTime) {
-              nestedCue.startTime = cueTime;
-            }
-            break;
-          }
-          case 'ruby':
-          case 'rp':
-          case 'rt':
-            nestedCue.rubyTag = tag;
-            break;
-          default:
-            break;
-        }
-      }
-    }
-
-    const isTextNode = (item) => shaka.util.TXml.isText(item);
-    const childNodes = element.children;
-    if (isTextNode(element) ||
-        (childNodes.length == 1 && isTextNode(childNodes[0]))) {
-      // Trailing line breaks may lost when convert cue to HTML tag
-      // Need to insert line break cue to preserve line breaks
-      const textArr = TXml.getTextContents(element).split('\n');
-      let isFirst = true;
-      for (const text of textArr) {
-        if (!isFirst) {
-          const lineBreakCue = shaka.text.Cue.lineBreak(
-              nestedCue.startTime, nestedCue.endTime);
-          rootCue.nestedCues.push(lineBreakCue);
-        }
-        if (text.length > 0) {
-          const textCue = nestedCue.clone();
-          textCue.payload = shaka.util.StringUtils.htmlUnescape(text);
-          rootCue.nestedCues.push(textCue);
-        }
-        isFirst = false;
-      }
-    } else {
-      rootCue.nestedCues.push(nestedCue);
-      for (const childNode of childNodes) {
-        VttTextParser.generateCueFromElement_(childNode, nestedCue, styles);
-      }
-    }
   }
 
   /**
@@ -974,31 +604,6 @@ shaka.text.VttTextParser = class {
 
     return true;
   }
-
-  /**
-   * Parses a WebVTT time from the given parser.
-   *
-   * @param {!shaka.util.TextParser} parser
-   * @return {?number}
-   * @private
-   */
-  static parseTime_(parser) {
-    const results = parser.readRegex(shaka.text.VttTextParser.timeFormat_);
-    if (results == null) {
-      return null;
-    }
-    // This capture is optional, but will still be in the array as undefined,
-    // in which case it is 0.
-    const hours = Number(results[1]) || 0;
-    const minutes = Number(results[2]);
-    const seconds = Number(results[3]);
-    const milliseconds = Number(results[4]);
-    if (minutes > 59 || seconds > 59) {
-      return null;
-    }
-
-    return (milliseconds / 1000) + seconds + (minutes * 60) + (hours * 3600);
-  }
 };
 
 /**
@@ -1006,15 +611,6 @@ shaka.text.VttTextParser = class {
  * @private
  */
 shaka.text.VttTextParser.MPEG_TIMESCALE_ = 90000;
-
-/**
- * @const
- * @private {!RegExp}
- * @example 00:00.000 or 00:00:00.000 or 0:00:00.000 or
- * 00:00.00 or 00:00:00.00 or 0:00:00.00
- */
-shaka.text.VttTextParser.timeFormat_ =
-    /(?:(\d{1,}):)?(\d{2}):(\d{2})\.(\d{2,3})/g;
 
 shaka.text.TextEngine.registerParser(
     'text/vtt', () => new shaka.text.VttTextParser());
